+++ src/aks-preview/HISTORY.md
@@ -3,9 +3,15 @@
 Release History
 ===============
 
-0.5.54
+0.5.55
 +++++
 * Add option `none` to `--network-plugin` parameter to skip CNI installation during cluster creation.
+
+0.5.54
+++++++
+* Add --host-group-id to `az aks create` and `az aks nodepool add` commands to support Azure Dedicated Host Group, which requires registering the feature flag "Microsoft.ContainerService/DedicatedHostGroupPreview"
+  * `az aks create --host-group-id`
+  * `az aks nodepool add --host-group-id`
 
 0.5.53
 ++++++